created template class and some unit testing
created the stack function along with the + - /  * and negate functions. tested unit testing functions such as assertequals aresame isnull etc. and tested to see if it would come up with an error for one input. still need to fix the maths behind it to print around the correct order such as the add and subtract in the tests.